fix #5 - roster add/remove buttons
the RosterPanel is a @Component so that it received RosterChangeEvents, but MainWindow was constructing an instance directly, resulting in two RosterPanels -- one that user interacts with on screen, and one that received changes from RosterService. MainWindow now uses the Autowired RosterPanel so that everything is using the same RosterPanel instance.